+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to lara-service will be documented in this file
 
-## 1.0.0 - 2018-05-21
+## 0.1.0 - 2018-05-21
 
 - initial release
+
+## 0.2.0 - 2018-06-05
+
+- Add autoloading of services
+- Add definition and handler suffixes to configuration
+- Add 'autoload' option to configuration
+- Fix Handler naming issue when using make:service command